requirements.txt
# Google API & Authentication
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0

# Configuration & Environment
python-dotenv==1.0.0

# HTTP & Networking
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22
chardet==4.0.0
idna==3.4

# Security & Encryption
keyring>=24.0.0
cryptography>=41.0.0
pysqlcipher3>=1.2.0

# Machine Learning & NLP
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
shap>=0.42.0
imbalanced-learn>=0.11.0
numpy>=1.24.0
pandas>=2.0.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Database
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
responses>=0.23.0

# CLI & UI
tqdm>=4.66.0
colorama>=0.4.6
click>=8.1.0

# Utilities
joblib>=1.3.0 